requirements.txt
pandas
numpy
dill
scikit-learn
python-dotenv
flask
certifi
pymongo[srv]
pymongo
pyaml
mlflow
dagshub
-e .